+++ example/lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:example/potatoo.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:input_quantity/input_quantity.dart';
 
@@ -36,34 +36,138 @@ class _MyHomePageState extends State<MyHomePage> {
       appBar: AppBar(
         title: Text(widget.title),
       ),
-      body: Center(
-        child: Padding(
-          padding: const EdgeInsets.all(18.0),
+      body: Padding(
+        padding: const EdgeInsets.all(18.0),
+        child: SingleChildScrollView(
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             crossAxisAlignment: CrossAxisAlignment.center,
             children: [
-              /// Input quantity widget
+              const Padding(
+                padding: EdgeInsets.all(8.0),
+                child: Text(
+                  'INPUT QUANTITY',
+                  style: TextStyle(fontSize: 24, fontWeight: FontWeight.w500),
+                ),
+              ),
+              const Padding(
+                padding: EdgeInsets.all(4.0),
+                child: Text(
+                  'by: pmatatias.dev',
+                  style: TextStyle(fontSize: 14, color: Colors.blue),
+                ),
+              ),
+              const Divider(),
+              const Padding(
+                padding: EdgeInsets.all(8.0),
+                child: Text('Easy and solid'),
+              ),
+              InputQty(
+                maxVal: 5,
+                initVal: 0,
+                steps: 1,
+                minVal: -5,
+                borderShape: BorderShapeBtn.circle,
+                boxDecoration: const BoxDecoration(),
+                minusBtn: const Icon(
+                  Icons.accessibility_new_outlined,
+                  size: 18,
+                ),
+                plusBtn: const Icon(
+                  Icons.accessible_forward,
+                  size: 18,
+                ),
+                btnColor1: Colors.indigo,
+                onQtyChanged: (val) {
+                  if (kDebugMode) {
+                    print(val);
+                  }
+                },
+              ),
+              const Divider(),
+              const Padding(
+                padding: EdgeInsets.all(8.0),
+                child: Text('With steps : 10 (customizable)'),
+              ),
               InputQty(
                 maxVal: 100,
                 initVal: 0.0,
                 steps: 10,
                 minVal: -100,
                 onQtyChanged: (val) {
-                  print(val);
+                  if (kDebugMode) {
+                    print(val);
+                  }
                 },
               ),
-              const SizedBox(
-                height: 30,
+              const Divider(),
+              const Padding(
+                padding: EdgeInsets.all(8.0),
+                child: Text('Min,max and initial value. Change color'),
               ),
-              TextButton(
-                  onPressed: () {
-                    Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => const PreviewExample()));
-                  },
-                  child: const Text('Preview'))
+              InputQty(
+                maxVal: 50,
+                initVal: 12,
+                minVal: -50,
+                steps: 10,
+                borderShape: BorderShapeBtn.none,
+                plusBtn: const Icon(Icons.add_box),
+                minusBtn: const Icon(Icons.indeterminate_check_box),
+                btnColor1: Colors.teal,
+                btnColor2: Colors.red,
+                onQtyChanged: (val) {
+                  if (kDebugMode) {
+                    print(val);
+                  }
+                },
+              ),
+              const Divider(),
+              const Padding(
+                padding: EdgeInsets.all(8.0),
+                child: Text('Intrinsic width ( default )'),
+              ),
+              InputQty(
+                maxVal: 100,
+                initVal: 0,
+                onQtyChanged: (val) {
+                  if (kDebugMode) {
+                    print(val);
+                  }
+                },
+              ),
+              const Divider(),
+              const Padding(
+                padding: EdgeInsets.all(8.0),
+                child: Text('Expanded width'),
+              ),
+              InputQty(
+                maxVal: 100,
+                initVal: 0,
+                isIntrinsicWidth: false,
+                onQtyChanged: (val) {
+                  if (kDebugMode) {
+                    print(val);
+                  }
+                },
+              ),
+              const Divider(),
+              const Padding(
+                padding: EdgeInsets.all(8.0),
+                child: Text('Customize input size'),
+              ),
+              InputQty(
+                maxVal: 100.0,
+                initVal: 0.0,
+                textFieldDecoration: const InputDecoration(
+                    isDense: false,
+                    contentPadding: EdgeInsets.symmetric(horizontal: 10)),
+                minVal: -100.0,
+                onQtyChanged: (val) {
+                  if (kDebugMode) {
+                    print(val);
+                  }
+                },
+              ),
             ],
           ),
         ),